Example of using bumpver to increment version strings
Instead of importing and dealing with eventual cyclical imports, this
uses bumpver to increment these values everywhere.

We've used bumpversion/bump2version in the past, and it's a bit more
opinionated and hard to configure. This is a much nicer tool. I'll be
raising switching our use to this tool on more repos.